refactor: APIProtocol 기능 분리
