+++ OpenMarket/OpenMarket/Networking/API/APIProtocol/APIProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  APIProtocol.swift
+//  OpenMarket
+//
+//  Created by 전민수 on 2022/07/26.
+//
+
+import Foundation
+
+protocol APIProtocol {
+    var configuration: APIConfiguration { get }
+}
